requirements.txt
async-timeout==4.0.2
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
packaging==21.3
Pillow==9.3.0
Pygments==2.13.0
pyparsing==3.0.9
redis==4.3.5
rich==12.6.0
shellingham==1.5.0
typer==0.7.0
walrus==0.9.2
